Defined DTOs

+++ Models/Transactions/Transaction.cs
@@ -5,5 +5,5 @@ public class Transaction
     public required int TransactionId { get; set; }
     public required int PayerId { get; set; }
     public required int ReceiverId { get; set; }
-    public required int Amount { get; set; }
+    public required double Amount { get; set; }
 }
